+++ src/db/rocksdb/rocksdb_engine.cpp
@@ -221,7 +221,7 @@ OP_NAMESPACE_BEGIN
                     int n = vsnprintf(buffer, buf_len - 1, format, ap);
                     if (n > 0 && NULL != buffer)
                     {
-                        if (n < buf_len)
+                        if ((size_t)n < buf_len)
                         {
                             buffer[n] = 0;
                         }
@@ -417,7 +417,7 @@ OP_NAMESPACE_BEGIN
                     {
                         return false;
                     }
-                    if (meta.GetTTL() > 0 && meta.GetTTL() <= get_current_epoch_millis())
+                    if (meta.GetTTL() > 0 && meta.GetTTL() <= (int64_t)get_current_epoch_millis())
                     {
                         g_db->AddExpiredKey(ns, k.GetKey());
                         return false;
